+++ src/Rules/Accessibility/ImagesHaveAltAttributes.php
@@ -14,7 +14,7 @@ class ImagesHaveAltAttributes extends BaseRule
     public function __construct()
     {
         $this->name = "imagesHaveAltAttributes";
-        $this->value = 12;
+        $this->value = 14;
         $this->type = RuleType::Accessibility;
     }