+++ src/Rules/Accessibility/MetaViewportPresent.php
@@ -14,7 +14,7 @@ class MetaViewportPresent extends BaseRule
     public function __construct()
     {
         $this->name = "metaViewportPresent";
-        $this->value = 12;
+        $this->value = 14;
         $this->type = RuleType::Accessibility;
     }