+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- When all accessibility rules pass, the total is now 100 as expected ([#4](https://github.com/lightship-core/lightship-php/issues/4)).
+
 ## [0.7.0] 2022-05-04
 
 ### Added